Move QR bitmap generation to a util class
